+++ src/contactForm.js
@@ -77,6 +77,14 @@ class RecrasContactForm {
         return this.element.querySelectorAll(querystring);
     }
 
+    isStandalone(options) {
+        if (options.showSubmit) {
+            console.warn('Option "showSubmit" was renamed to "standalone". Please update your code.');
+            options.standalone = true;
+        }
+        return !!options.standalone;
+    }
+
     generateForm(extraOptions = {}) {
         let waitFor = [];
 
@@ -88,14 +96,16 @@ class RecrasContactForm {
             RecrasCSSHelper.loadCSS('pikaday');
         }
         return Promise.all(waitFor).then(() => {
-            let html = '<form class="recras-contactform">';
+            const standalone = this.isStandalone(extraOptions);
+            const validateText = standalone ? 'novalidate' : '';
+            let html = `<form class="recras-contactform" ${ validateText }>`;
             if (extraOptions.voucherQuantitySelector) {
                 html += this.quantitySelector();
             }
             this.contactFormFields.forEach((field, idx) => {
                 html += '<div>' + this.showField(field, idx) + '</div>';
             });
-            if (extraOptions.showSubmit) {
+            if (standalone) {
                 html += this.submitButtonHtml();
             }
             html += '</form>';
@@ -118,6 +128,9 @@ class RecrasContactForm {
             }
             contactForm[field.dataset.identifier].push(field.value);
         });
+        if (contactForm['boeking.datum']) {
+            contactForm['boeking.datum'] = RecrasDateHelper.formatStringForAPI(contactForm['boeking.datum']);
+        }
 
         return contactForm;
     }
@@ -160,6 +173,30 @@ class RecrasContactForm {
         });
     }
 
+    getInvalidFields() {
+        let invalid = [];
+        let required = this.getEmptyRequiredFields();
+
+        let els = this.findElements('.recras-contactform :invalid');
+        for (let el of els) {
+            if (!required.includes(el)) {
+                invalid.push(el);
+            }
+        }
+        return invalid;
+    }
+
+    getEmptyRequiredFields() {
+        let isEmpty = [];
+        let els = this.findElements('.recras-contactform :required');
+        for (let el of els) {
+            if (el.value === undefined || el.value === '') {
+                isEmpty.push(el);
+            }
+        }
+        return isEmpty;
+    }
+
     hasFieldOfType(identifier) {
         return this.contactFormFields.filter(field => {
             return field.field_identifier === identifier;
@@ -173,6 +210,16 @@ class RecrasContactForm {
     }
     hasPackageField() {
         return this.hasFieldOfType('boeking.arrangement');
+    }
+
+    isEmpty() {
+        let els = this.findElements('.recras-contactform input, .recras-contactform select, .recras-contactform textarea');
+        let formValues = [...els].map(el => el.value);
+        return !formValues.some(v => v !== '');
+    }
+
+    isValid() {
+        return this.findElement('.recras-contactform').checkValidity();
     }
 
     loadingIndicatorHide() {
@@ -192,10 +239,20 @@ class RecrasContactForm {
         return `<div><label for="number-of-vouchers">${ this.languageHelper.translate('VOUCHER_QUANTITY') }</label><input type="number" id="number-of-vouchers" class="number-of-vouchers" min="1" value="1" required></div>`;
     }
 
+    removeErrors(parentQuery = '') {
+        [...this.findElements(parentQuery + ' .booking-error')].forEach(el => {
+            el.parentNode.removeChild(el);
+        });
+    }
+
     removeWarnings() {
         [...this.findElements('.recrasError')].forEach(el => {
             el.parentNode.removeChild(el);
         });
+    }
+
+    hasEmptyRequiredFields() {
+        return this.getEmptyRequiredFields().length > 0;
     }
 
     showField(field, idx) {
@@ -204,7 +261,6 @@ class RecrasContactForm {
         }
 
         const today = RecrasDateHelper.toString(new Date());
-        const datePattern = '[0-9]{4}-(0[1-9]|1[012])-(0[1-9]|1[0-9]|2[0-9]|3[01])';
         const timePattern = '(0[0-9]|1[0-9]|2[0-3])(:[0-5][0-9])';
 
         let label = this.showLabel(field, idx);
@@ -271,7 +327,7 @@ class RecrasContactForm {
                 return label + html;
             case 'boeking.datum':
                 placeholder = this.languageHelper.translate('DATE_FORMAT');
-                return label + `<input type="text" ${ fixedAttributes } min="${ today }" placeholder="${ placeholder }" pattern="${ datePattern }" autocomplete="off">`;
+                return label + `<input type="text" ${ fixedAttributes } min="${ today }" placeholder="${ placeholder }" autocomplete="off">`;
             case 'boeking.groepsgrootte':
                 return label + `<input type="number" ${ fixedAttributes } min="1">`;
             case 'boeking.starttijd':
@@ -279,6 +335,13 @@ class RecrasContactForm {
                 return label + `<input type="time" ${ fixedAttributes } placeholder="${ placeholder }" pattern="${ timePattern }" step="300">`;
             case 'boeking.arrangement':
                 const preFilledPackage = this.options.getPackageId();
+                if (field.verplicht && this.packages.length === 1) {
+                    let pack = this.packages[0];
+                    html = `<select ${ fixedAttributes }>
+                        <option value="${ pack.id }" selected>${ pack.arrangement }
+                    </select>`;
+                    return label + html;
+                }
 
                 html = `<select ${ fixedAttributes }>`;
                 html += `<option value="">`;
@@ -300,7 +363,7 @@ class RecrasContactForm {
         this.loadingIndicatorShow(this.element);
         return this.getContactFormFields()
             .then(() => this.generateForm({
-                showSubmit: true,
+                standalone: true,
             }))
             .then(html => {
                 this.appendHtml(html);
@@ -312,9 +375,6 @@ class RecrasContactForm {
                             field: this.findElement('[data-identifier="boeking.datum"]'),
                             i18n: RecrasCalendarHelper.i18n(this.languageHelper),
                             numberOfMonths: 1,
-                            toString(date, _format) {
-                                return RecrasDateHelper.datePartOnly(date);
-                            },
                         }
                     );
 
@@ -322,6 +382,29 @@ class RecrasContactForm {
                 }
                 this.loadingIndicatorHide();
             });
+    }
+
+    showInlineErrors() {
+        for (let el of this.getEmptyRequiredFields()) {
+            const labelEl = el.parentNode.querySelector('label');
+            const requiredText = this.languageHelper.translate('CONTACT_FORM_FIELD_REQUIRED', {
+                FIELD_NAME: labelEl.innerText,
+            });
+            el.parentNode.insertAdjacentHTML(
+                'afterend',
+                `<div class="booking-error">${ requiredText }</div>`
+            );
+        }
+        for (let el of this.getInvalidFields()) {
+            const labelEl = el.parentNode.querySelector('label');
+            const invalidText = this.languageHelper.translate('CONTACT_FORM_FIELD_INVALID', {
+                FIELD_NAME: labelEl.innerText,
+            });
+            el.parentNode.insertAdjacentHTML(
+                'afterend',
+                `<div class="booking-error">${ invalidText }</div>`
+            );
+        }
     }
 
     showLabel(field, idx) {
@@ -338,13 +421,26 @@ class RecrasContactForm {
 
     submitForm(e) {
         e.preventDefault();
-        this.eventHelper.sendEvent(RecrasEventHelper.PREFIX_CONTACT_FORM, RecrasEventHelper.EVENT_CONTACT_FORM_SUBMIT, this.options.getFormId());
+
         let submitButton = this.findElement('.submitForm');
 
-        let status = this.checkRequiredCheckboxes();
-        if (!status) {
+        this.removeErrors('.recras-contactform');
+        if (this.isEmpty()) {
+            submitButton.parentNode.insertAdjacentHTML(
+                'afterend',
+                `<div class="booking-error">${ this.languageHelper.translate('ERR_CONTACT_FORM_EMPTY') }</div>`
+            );
+            return false;
+        } else if (this.hasEmptyRequiredFields() || !this.isValid()) {
+            this.showInlineErrors();
             return false;
         }
+
+        if (!this.checkRequiredCheckboxes()) {
+            return false;
+        }
+
+        this.eventHelper.sendEvent(RecrasEventHelper.PREFIX_CONTACT_FORM, RecrasEventHelper.EVENT_CONTACT_FORM_SUBMIT, this.options.getFormId());
 
         this.loadingIndicatorHide();
         this.loadingIndicatorShow(submitButton);