+++ test/js-unit/recrasBookingSpec.js
@@ -322,4 +322,186 @@ describe('RecrasBooking', () => {
             }));
         });
     });
+
+    describe('Option "package_id"', () => {
+        let rb;
+        const package1 = {
+            id: 1,
+        };
+        const package2 = {
+            id: 2,
+        };
+        const package3 = {
+            id: 3,
+        };
+        const package4 = {
+            id: 4,
+        };
+        const getPackagesMock = () => [package1, package2, package3, package4];
+
+        it('works with a single package', async () => {
+            rb = new RecrasBooking(new RecrasOptions({
+                element: document.createElement('div'),
+                recras_hostname: 'demo.recras.nl',
+                package_id: 2,
+            }));
+            spyOn(rb, 'showPackages');
+            spyOn(rb, 'changePackage');
+            rb.getPackages = getPackagesMock;
+
+            await rb.promise;
+
+            expect(rb.changePackage).toHaveBeenCalledWith(2);
+            expect(rb.showPackages).not.toHaveBeenCalled();
+        });
+
+        it('works with a single-item array', async () => {
+            rb = new RecrasBooking(new RecrasOptions({
+                element: document.createElement('div'),
+                recras_hostname: 'demo.recras.nl',
+                package_id: [2],
+            }));
+            spyOn(rb, 'showPackages');
+            spyOn(rb, 'changePackage');
+            rb.getPackages = getPackagesMock;
+
+            await rb.promise;
+
+            expect(rb.changePackage).toHaveBeenCalledWith(2);
+            expect(rb.showPackages).not.toHaveBeenCalled();
+        });
+
+        it('filters the visible packages', async () => {
+            rb = new RecrasBooking(new RecrasOptions({
+                element: document.createElement('div'),
+                recras_hostname: 'demo.recras.nl',
+                package_id: [2, 3],
+            }));
+            spyOn(rb, 'showPackages');
+            rb.getPackages = getPackagesMock;
+
+            await rb.promise;
+
+            expect(rb.showPackages).toHaveBeenCalledWith([package2, package3]);
+        });
+
+        it('works without option', async () => {
+            rb = new RecrasBooking(new RecrasOptions({
+                element: document.createElement('div'),
+                recras_hostname: 'demo.recras.nl',
+            }));
+            spyOn(rb, 'showPackages');
+            rb.getPackages = getPackagesMock;
+
+            await rb.promise;
+
+            expect(rb.showPackages).toHaveBeenCalledWith([package1, package2, package3, package4]);
+        });
+    });
+
+    describe('requiredAmount', () => {
+        let rb;
+        beforeEach(() => {
+            rb = new RecrasBooking(new RecrasOptions({
+                element: document.createElement('div'),
+                recras_hostname: 'demo.recras.nl',
+            }));
+        });
+
+        it('should calculate "needs 3 per 2, rounded up"', () =>
+            expect(rb.requiredAmount(1, {
+                aantal: 3,
+                per_x_aantal: 2,
+                afronding: 'boven',
+            })).toEqual(3)
+        );
+
+        it('should calculate "needs 4 per 2, rounded down"', () =>
+            expect(rb.requiredAmount(3, {
+                aantal: 4,
+                per_x_aantal: 2,
+                afronding: 'beneden',
+            })).toEqual(4)
+        );
+    });
+
+    describe('checkMinMaxAmounts', () => {
+        let rb;
+
+        beforeEach(() => {
+            let mainEl = document.createElement('div');
+            mainEl.id = 'onlinebooking';
+            document.body.appendChild(mainEl);
+            rb = new RecrasBooking(new RecrasOptions({
+                element: mainEl,
+                recras_hostname: 'demo.recras.nl',
+                package_id: 26,
+            }));
+            spyOn(rb, 'setMinMaxAmountWarning');
+
+            // Mocks
+            rb.getSetting = (name) => ({
+                slug: name,
+                waarde: '',
+            });
+            rb.getPackages = () => {
+                const packages = [{
+                    id: 26,
+                    onlineboeking_contactformulier_id: 4,
+                    regels: [
+                        {
+                            id: 669,
+                            max: 8,
+                            onlineboeking_aantalbepalingsmethode: "invullen_door_gebruiker",
+                            product: {
+                                minimum_aantal: 5,
+                                vereist_product: [],
+                            }
+                        },
+                        {
+                            id: 1337,
+                            max: null,
+                            onlineboeking_aantalbepalingsmethode: "invullen_door_gebruiker",
+                            product: {
+                                minimum_aantal: 1,
+                                vereist_product: [],
+                            }
+                        },
+                    ],
+                }];
+                rb.packages = packages;
+                return packages;
+            };
+        });
+
+        it('gives error if amount is less than the minimum', async () => {
+            await rb.promise;
+
+            let el = document.getElementById('packageline0');
+            el.value = '1';
+            el.dispatchEvent(new Event('input'));
+
+            expect(rb.setMinMaxAmountWarning).toHaveBeenCalledWith('packageline0', 5, 'minimum');
+        });
+
+        it('gives error if amount is more than the maximum', async () => {
+            await rb.promise;
+
+            let el = document.getElementById('packageline0');
+            el.value = '15';
+            el.dispatchEvent(new Event('input'));
+
+            expect(rb.setMinMaxAmountWarning).toHaveBeenCalledWith('packageline0', 8, 'maximum');
+        });
+
+        it('does not give error if there is no maximum', async () => {
+            await rb.promise;
+
+            let el = document.getElementById('packageline1');
+            el.value = '2';
+            el.dispatchEvent(new Event('input'));
+
+            expect(rb.setMinMaxAmountWarning).not.toHaveBeenCalled();
+        });
+    });
 });